refactor(checker): split diagnostic source helpers
